requirements.txt
# Core Machine Learning & Deep Learning
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Natural Language Processing
sentence-transformers>=2.2.2
spacy>=3.6.0
huggingface-hub>=0.16.0

# Vector Search & Similarity
faiss-gpu>=1.7.4  # Use faiss-cpu if no GPU available

# Graph Processing
networkx>=3.1

# Data Processing & Utilities
ujson>=5.8.0
tqdm>=4.65.0
joblib>=1.3.0

# Visualization & Plotting
matplotlib>=3.7.1
Pillow>=10.0.0

# Development & Utilities
jupyter>=1.0.0
ipykernel>=6.25.0
peft
torch-scatter
wandb
ujson
# Optional: For distributed training
# torch-tb-profiler>=0.4.0  # Uncomment if using TensorBoard profiling 